requirements.txt
Flask
tabula-py
pandas
openpyxl
waitress
